Migrated the project over to use the modern .vue file format
