+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "app": "node server.js",
     "app:dev": "nodemon server.js",
-    "test": "jest"
+    "test": "jest",
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build"
   },
   "repository": {
     "type": "git",
@@ -19,19 +21,32 @@
   },
   "homepage": "https://github.com/coltonhyer/ApartmentPal#readme",
   "dependencies": {
+    "@vue/cli": "^5.0.4",
     "axios": "^0.26.0",
+    "core-js": "^3.8.3",
+    "cors": "^2.8.5",
     "express": "^4.17.3",
     "helmet": "^5.0.2",
     "mongodb": "^4.4.1",
     "mongoose": "^6.2.4",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.15",
+    "path": "^0.12.7",
     "serve-favicon": "^2.5.0",
     "sha.js": "^2.4.11",
-    "vue": "^3.2.31",
+    "vue": "^2.6.14",
+    "vue-router": "^3.5.1",
     "vuetify": "^2.6.4"
   },
   "devDependencies": {
-    "jest": "^27.5.1"
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-router": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "jest": "^27.5.1",
+    "sass": "~1.32.0",
+    "sass-loader": "^10.0.0",
+    "vue-cli-plugin-vuetify": "~2.4.8",
+    "vue-template-compiler": "^2.6.14",
+    "vuetify-loader": "^1.7.0"
   }
 }
