--- public/styles/index.css
@@ -1,91 +1,0 @@
-html, body{
-    height:100% !important;
-    width:100%;
-    margin:0;
-    padding:0;
-}
-#app{
-    height:inherit;
-}
- 
-/*Styles for the login page*/
-.login{
-    height:100%;
-}
-
-#welcome{
-    height: 30%;
-    background-color: #4A5240;
-    display:flex;
-    margin:0;
-    align-items: center;
-    justify-content: center;
-}
-#welcome div{
-    font-size:100px;
-    font-family: sans-serif;
-    color:#a28497;
-}
-
-#loginMessage{
-    height:10%;
-    display:flex;
-    align-items:center;
-    justify-content: center;
-    background-color: #ADA8BE;
-}
-
-#loginMessage{
-    font-size: 30px;
-    font-family:sans-serif;
-}
-
-#loginContainer{
-    display:flex;
-    height:55%;
-    align-items:center;
-    justify-content:center;
-}
-
-#loginForm{
-    display:flex;
-    flex-direction:column;
-    bottom:75px;
-    justify-content:center;
-}
-
-/* Styles for the admin views*/
-.adminBanner{
-    height:10%;
-    background-color: #6f5e5c;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.adminBanner div{
-    font-size:50px;
-    font-family:sans-serif;
-    color: #c6caed;
-}
-
-/* Lookup Page */
-.attendantLookup{
-    height:100%;
-}
-
-#lookupField{
-    display:flex;
-    height:90%;
-    justify-content: center;
-    align-items: center;
-}
-
-#lookupForm{
-    display:flex;
-    flex-direction:column;
-    align-items:center;
-    width: 100%;
-}
-
-
