+++ src/App.vue
@@ -1,0 +1,19 @@
+<template>
+  <v-app>
+      <v-main>
+          <router-view></router-view>
+      </v-main>
+  </v-app>
+</template>
+
+<script>
+
+export default {
+  name: 'App',
+
+  data: () => ({
+    passInfo: null,
+    residentID: null,
+  }),
+};
+</script>
