+++ src/plugins/vuetify.js
@@ -1,0 +1,6 @@
+import Vue from 'vue';
+import Vuetify from 'vuetify/lib/framework';
+
+Vue.use(Vuetify);
+
+export default new Vuetify();
