+++ src/router/views.js
@@ -1,0 +1,14 @@
+import Login from '../views/Login.vue'
+import Lookup from '../views/Lookup.vue'
+import Results from '../views/Results.vue'
+import Passes from '../views/Passes.vue'
+import PassForm from '../views/PassForm.vue'
+
+
+export default [
+    Login,
+    Lookup,
+    Results,
+    Passes,
+    PassForm
+]
