+++ vue.config.js
@@ -1,0 +1,9 @@
+const { defineConfig } = require('@vue/cli-service')
+module.exports = defineConfig({
+  transpileDependencies: [
+    'vuetify'
+  ],
+  devServer: {
+    proxy: 'http://localhost:3000'
+  }
+})
